version 0.5

+++ api/routes/os.js
@@ -2,6 +2,12 @@ const express = require('express');
 const router = express.Router();
 
 const fs = require('fs');
+const j2t = require('json2toml');
+const { Readable } = require('stream');
+const { finished } = require('stream/promises');
+
+const modrinth = require('../utils/modrinth');
+const curseforge = require('../utils/curseforge');
 
 router.get('/', (req, res, next) => {
     res.send('ping');
@@ -12,7 +18,7 @@ router.post('/files', (req, res, next) => {
 
     if (!path) {
         res.statusCode = 400;
-        res.json({error: 'Missing path'});
+        res.json({ error: 'Missing path' });
         return;
     }
 
@@ -27,19 +33,64 @@ router.post('/files', (req, res, next) => {
     });
 });
 
-router.post('/download', (req, res, next) => {
-    const { path, download } = req.body;
-});
+router.post('/download', async (req, res, next) => {
+    const { path, modinfo, gameversion, loaders } = req.body;
 
-router.post('/delete', (req, res, next) => {
-    const { path, filename } = req.body;
-
-    if (!path) {
+    if (!path || !modinfo) {
         res.statusCode = 400;
-        res.json({error: 'Missing path'});
+        res.json({ error: 'Bad Request' });
         return;
     }
 
+    let data;
+    const pId = modinfo.url.split('/').at(-1);
+    if (modinfo.url.search('modrinth') > 0) {
+        data = await modrinth.getInfo(
+            pId,
+            modinfo.filename,
+            modinfo.version,
+            gameversion,
+            loaders
+        );
+    }
+    if (modinfo.url.search('curseforge') > 0) {
+        data = await curseforge.getInfo(
+            pId,
+            modinfo.filename,
+            modinfo.version,
+            gameversion,
+            loaders
+        );
+    }
+
+    if (!data) {
+        res.statusCode = 400;
+        res.json({ error: `${modinfo.filename} failed`});
+        return;
+    }
+
+    const filestream = fs.createWriteStream(`${path}/${data.filename}`);
+    const download = await fetch(data.download.url, {method: 'GET'});
+    finished(Readable.fromWeb(download.body).pipe(filestream));
+
+    filestream.on('finish', () => {
+        filestream.close();
+        fs.writeFileSync(`${path}/.index/${data.slug}.pw.toml`, j2t(data, {newlineAfterSection: true}));
+        res.json({ msg: 'Success', data: data });
+    })
+
+});
+
+router.post('/delete', (req, res, next) => {
+    const { path, filename, tomlname } = req.body;
+
+    if (!path || !filename) {
+        res.statusCode = 400;
+        res.json({error: 'Bad Request'});
+        return;
+    }
+
+    // this is so stupid but i stopped caring
     fs.rm(`${path}/${filename}`, (err) => {
         if (err) {
             res.statusCode = 400;
@@ -47,7 +98,16 @@ router.post('/delete', (req, res, next) => {
             return;
         }
 
-        res.json({ msg: 'Success' });
+        fs.rm(`${path}/.index/${tomlname}`, (err) => {
+            if (err) {
+                res.statusCode = 400;
+                res.json(err);
+                return;
+            }
+
+            res.json({ msg: 'Success' });
+        })
+
     })
 });
